DCS-1166 extracting out conditions to new location
